Fix relative location of the start script
The compilation of a very simple example with a start
script in the main folder was failing (see #399)
because the `start` script was not found.
Most probably, the reason is that the folder in which
the script resides is not in the PATH. To solve this,
we replace the `start` string with a relative `./start`,
which solves the problem and completes the build and the
run.