Remove underlines from H2 headers that are links
The `contents::` directive in Sphinx causes the headers of referred
sections to become links back to the line in the table of
contents. This causes the headers to use the default attributes for
links, including underlines. The latter makes the header style out of
line with the rest of the documentation. A single `text-decoration:
none` in the custom CSS file fixes this (note: only for H2 headers
currently: these were the offending ones).