Add Pipfile + Pipfile.lock BuildPack logic
This is the actual meat while most previous commits were tests to verify
that this worked as intented. The tests paid off quickly as this was
quite hard to get working at all.

- We can not use `pipenv shell` inside this script.
- We must use `--system` and therefore `pipenv install` that supports it
rather than `pipenv sync` that doesn.
- We must generate a `Pipfile.lock` if there are none.
- We must use `--ignore-pipfile` and `--deploy` on the `pipenv install`
command in order to mimic how a `pipenv sync` would have worked.